refactor: Voucher 추상 클래스로 리팩터링
