+++ src/main/java/com/programmers/vouchermanagement/domain/voucher/FixedAmountVoucher.java
@@ -1,23 +1,11 @@
 package com.programmers.vouchermanagement.domain.voucher;
 
 import lombok.Getter;
-import lombok.ToString;
-
-import java.util.UUID;
 
 @Getter
-@ToString
-public class FixedAmountVoucher implements Voucher {
-    private UUID id;
-    private final long amount;
-
+public class FixedAmountVoucher extends Voucher {
     public FixedAmountVoucher(long amount) {
-        this.amount = amount;
-    }
-
-    @Override
-    public void setId(UUID id) {
-        this.id = id;
+        super(amount);
     }
 
     @Override
