+++ src/main/java/com/programmers/vouchermanagement/domain/voucher/Voucher.java
@@ -1,11 +1,26 @@
 package com.programmers.vouchermanagement.domain.voucher;
 
+import lombok.Getter;
+
 import java.util.UUID;
 
-public interface Voucher {
-    void setId(UUID id);
-    UUID getId();
-    long getAmount();
-    VoucherType getType();
-    String toString();
+@Getter
+public abstract class Voucher {
+    private UUID id;
+    protected final long amount;
+
+    public Voucher(long amount) {
+        this.amount = amount;
+    }
+
+    public void setId(UUID id) {
+        this.id = id;
+    }
+
+    public abstract VoucherType getType();
+
+    @Override
+    public String toString() {
+        return "Voucher{id=" + id + ", amount=" + amount + ", type=" + getType() + '}';
+    }
 }
