+++ src/main/java/com/programmers/vouchermanagement/repository/voucher/MemoryVoucherRepository.java
@@ -6,7 +6,10 @@ import com.programmers.vouchermanagement.util.UuidProvider;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Repository
